+++ main/ad2_uart_cli.h
@@ -25,7 +25,7 @@
 #define _AD2_UART_CLI_H_
 
 #define PROMPT_STRING "AD2IOT # "
-
+#define AD2_HELP_CMD "help"
 #define CLI_TASK_PRIORITY (tskIDLE_PRIORITY+2)
 #define CLI_TASK_SIZE    (8192)