+++ main/ad2_utils.h
@@ -52,6 +52,7 @@ std::string ad2_string_format(const std::string fmt, ...);
 int  ad2_copy_nth_arg(char* dest, char* src, int size, int n);
 void ad2_tokenize(std::string const &str, const char delim, std::vector<std::string> &out);
 std::string ad2_to_string(int n);
+int ad2_query_key_value(std::string &qry_str, const char * key, std::string &val);
 
 // NV Storage utilities
 
@@ -61,6 +62,9 @@ void ad2_set_nv_slot_key_int(const char *key, int slot, int value);
 void ad2_get_nv_slot_key_int(const char *key, int slot, int *value);
 void ad2_set_nv_slot_key_string(const char *key, int slot, char *value);
 void ad2_get_nv_slot_key_string(const char *key, int slot, char *value, size_t size);
+
+// misc
+char ad2_network_mode(std::string &args);
 
 #ifdef __cplusplus
 }